requirements.txt
isort
black
flake8
mypy
pyyaml
streamlit
google-cloud-aiplatform
python-dotenv
torch
transformers
accelerate